MINOR: Refactor the MetadataCache interface
Remove getNonExistingTopics, which was not necessary. MetadataCache
already lets callers check for the existence of topics by calling
MetadataCache#contains.

Add MetadataCache#getAliveBrokerNode and getAliveBrokerNodes.  This
simplifies the calling code, which always wants a Node.

Add ScalaDoc for MetadataCache#numPartitions.

MetadataCache#numPartitions should return a simple int rather than an
Option[Int]. A topic which exists can never have 0 partitions, so a
return of 0 is not ambiguous.

Fix a case where we were calling getAliveBrokers and filtering by id,
rather than simply calling getAliveBroker(id) and making use of the hash
map.

Put ZkMetadataCache into its own file, like the other MetadataCache
subclass.
