+++ core/src/main/scala/kafka/admin/AdminUtils.scala
@@ -100,7 +100,7 @@ object AdminUtils extends Logging {
    *                                 assign each replica to a unique rack.
    *
    */
-  def assignReplicasToBrokers(brokerMetadatas: Seq[BrokerMetadata],
+  def assignReplicasToBrokers(brokerMetadatas: Iterable[BrokerMetadata],
                               nPartitions: Int,
                               replicationFactor: Int,
                               fixedStartIndex: Int = -1,
@@ -124,7 +124,7 @@ object AdminUtils extends Logging {
 
   private def assignReplicasToBrokersRackUnaware(nPartitions: Int,
                                                  replicationFactor: Int,
-                                                 brokerList: Seq[Int],
+                                                 brokerList: Iterable[Int],
                                                  fixedStartIndex: Int,
                                                  startPartitionId: Int): Map[Int, Seq[Int]] = {
     val ret = mutable.Map[Int, Seq[Int]]()
@@ -147,7 +147,7 @@ object AdminUtils extends Logging {
 
   private def assignReplicasToBrokersRackAware(nPartitions: Int,
                                                replicationFactor: Int,
-                                               brokerMetadatas: Seq[BrokerMetadata],
+                                               brokerMetadatas: Iterable[BrokerMetadata],
                                                fixedStartIndex: Int,
                                                startPartitionId: Int): Map[Int, Seq[Int]] = {
     val brokerRackMap = brokerMetadatas.collect { case BrokerMetadata(id, Some(rack)) =>
