+++ core/src/main/scala/kafka/server/BrokerToControllerChannelManager.scala
@@ -73,8 +73,7 @@ class MetadataCacheControllerNodeProvider(
 ) extends ControllerNodeProvider {
   override def get(): Option[Node] = {
     metadataCache.getControllerId
-      .flatMap(metadataCache.getAliveBroker)
-      .map(_.endpoints(listenerName.value))
+      .flatMap(metadataCache.getAliveBrokerNode(_, listenerName.value))
   }
 }
