+++ core/src/test/scala/unit/kafka/server/ServerStartupTest.scala
@@ -86,9 +86,9 @@ class ServerStartupTest extends ZooKeeperTestHarness {
     val props = TestUtils.createBrokerConfig(brokerId, zkConnect)
     server = TestUtils.createServer(KafkaConfig.fromProps(props))
 
-    TestUtils.waitUntilTrue(() => server.metadataCache.getAliveBrokers.nonEmpty, "Wait for cache to update")
-    assertEquals(1, server.metadataCache.getAliveBrokers.size)
-    assertEquals(brokerId, server.metadataCache.getAliveBrokers.head.id)
+    TestUtils.waitUntilTrue(() => server.metadataCache.getAliveBrokers().nonEmpty, "Wait for cache to update")
+    assertEquals(1, server.metadataCache.getAliveBrokers().size)
+    assertEquals(brokerId, server.metadataCache.getAliveBrokers().head.id)
   }
 
   @Test
